Nexon Old Browser Security Update
As of 2/24/2021, Nexon changed security settings on their website(s), and now require IE version 10 or higher to function properly. IE versions 7, 8, and 9 no longer create the proper cookies requires for launch, and the WebBrowser class defaults to IE7. This change adjusts the FEATURE_BROWSER_EMULATION registry key and adds/updates an entry for GameStarter to use IE 10 or 11 (whichever is the lowest present on the host machine). If IE 9 or lower is the latest, it will error out and not allow the application to run. Updated app to version 1.3.